FortuneEngine reads json db
Added to the Fortune Engine object to have fortune_db, an object that keeps fortunes.
Added to the Fortune Engine object to have app_name, a string with the
appname.

Changed constructor to take argument of app_name.

Added a setter to add outcomes to keep outcomes functionality.

Added db_reader, a function that takes the .json file of fortune
tellings and adds them to FortuneEngine as an object.

Minor changements to engine.js arurrounding these.