Added some spice to fortune_sick
created+added css on h2 with id fortune-message that centers and sets a width to keep first message from changing.

Adopted TYPING_SPEED to actually change the speed of the typing.

Reworked function selectCategory(category)

Instead of using an html collection and making opacity 0 of all it first uses querySelector to select the cards that are not the one chosen. Only later is the collection called to make it selected.

  It uses a for loop in order to choose which were not chosen, and make them transparent. Though this also means that the one chosen is still there.

  Next goal is to move the chosen card to the center. Maybe after expand it to make it look like the text came from the other side but the fade still gets the idea accross.