Started adding selection cartomancy-tarot
Cards are mentioned to have their nonexistant id, would be better to use
a an array but for now just trying to remove depending on an Id.

Also started addinga  new container to move cards that are chosen to
that place.

Also brought some code from cartomancy branch just to read maybe not the
actual best to have since most of it is commented.