Added first 11 cards of tarot readings
Added some skeleton for the reading of the cards.

Also added the first 10 entries description in cartomancy.json

I note that it is missing the actual pictures in this. And some other
fields may be desirable. This is mostly just ot serve as a guide for the
next few steps.