Added a card and a description
For the time being it only does this with the first pickContainer.

It adds a stock image not of the card (pikachu) and has a desciption
text. Copied off https://codepen.io/keubibo/pen/yLXYVOK and that person
deserves a lot of the credit.

Scuffed because it adds removes the child of the container, the original
card, and replaces it with the new elements. Also to the container adds
a class that allows it to work with the description mentioned. In the
future it may be good to revise the css to just be the first css
class.