Changes to displaying the fortune
The fortunes are still not in the correct position. But the cards have
been resized so they show more of the lines needed.

Also inverted the color of the images to display better in a light theme
we have going on.

Also since there was a new div to hide the cards move the
pickContainer{1,2,3} out of there and new parent is <div id=center-div>.
Hopefully this will make a few things easier such as hiding the cards.