Added displaying the cards and their fortunes
Finally the fortune cards appear next to each other on a line as flex
objects.

Also deleted a header since we weren't using it, but we should have a
title someitme. Also added a class for correcting the container of pick
containers.
  Specifically since the containers are used after their css should
  change. So instead of referencing them by Id in the css the classes
  added are: pickContainer p1 and in the css are references as
  .pickContainer.p1
  Then later on after pressing the button we want to have a different
  format, so pickContainer class is removed and the new class,
  read-container, takes precedence.

Still some issues, like resizeing

Going to merge after this.